fix version 12 crash on receiving packet
