+++ src/enums.h
@@ -20,14 +20,6 @@
 #ifndef FS_ENUMS_H_003445999FEE4A67BCECBE918B0124CE
 #define FS_ENUMS_H_003445999FEE4A67BCECBE918B0124CE
 
-enum ConnectionState_t : int8_t {
-	CONNECTION_STATE_DISCONNECTED,
-	CONNECTION_STATE_CONNECTING_STAGE1,
-	CONNECTION_STATE_CONNECTING_STAGE2,
-	CONNECTION_STATE_GAME,
-	CONNECTION_STATE_PENDING
-};
-
 enum RuleViolationType_t : uint8_t {
 	REPORT_TYPE_NAME = 0,
 	REPORT_TYPE_STATEMENT = 1,
