+++ src/iologindata.cpp
@@ -104,9 +104,7 @@ bool IOLoginData::loginserverAuthentication(const std::string& name, const std::
 uint32_t IOLoginData::gameworldAuthentication(const std::string& accountName, const std::string& password, std::string& characterName /*, std::string& token, uint32_t tokenTime */)
 {
 	Database& db = Database::getInstance();
-	//std::cout << accountName << " " << password << " " << characterName << std::endl;
 	std::string q = fmt::format("SELECT `id`, `password`, `secret` FROM `accounts` WHERE `name` = {:s} OR `email` = {:s} LIMIT 1", db.escapeString(accountName), db.escapeString(accountName));
-	std::cout << q << std::endl;
 	DBResult_ptr result = db.storeQuery(q);
 	if (!result) {
 		return 0;
