+++ src/protocolold.cpp
@@ -68,7 +68,7 @@ void ProtocolOld::onRecvFirstMessage(NetworkMessage& msg)
 	setXTEAKey(std::move(key));
 
 	if (version <= 822) {
-		disableChecksum();
+		setChecksumMode(CHECKSUM_DISABLED);
 	}
 
 	disconnectClient(fmt::format("Only clients with protocol {:s} allowed!", CLIENT_VERSION_STR));
